Move resizable_state.dart to root directory.
