+++ patterns/state/manipulator_state/states/selections/resizable_state.dart
@@ -1,8 +1,8 @@
-import 'bottom_left_marker_state.dart';
-import 'bottom_right_marker_state.dart';
-import 'top_left_marker_state.dart';
-import 'top_right_marker_state.dart';
-import '../sub_states/parent_state.dart';
+import 'resizable_state/bottom_left_marker_state.dart';
+import 'resizable_state/bottom_right_marker_state.dart';
+import 'resizable_state/top_left_marker_state.dart';
+import 'resizable_state/top_right_marker_state.dart';
+import 'sub_states/parent_state.dart';
 
 class ResizableState extends ParentState {
   ResizableState({required super.selectedShape}) {
